Refactored types used by CarConfiguration
Use double rather than float.  Plays nicer with timespans
Using timespans to compare times in tests.
Wrote a diagnostic car configuration which logs out values making it
easy to see what values are set to in the test output.

+++ RaceCarSetup/RaceTrack.cs
@@ -6,7 +6,7 @@
 		{
 		}
 
-		public float LapDistance {
+		public int LapDistance {
 			get;
 			set;
 		}
@@ -16,7 +16,7 @@
 			set;
 		}
 
-		public float PitStopTime {
+		public int PitStopTime {
 			get;
 			set;
 		}
